AG-5781, AG-5786 - Pipeline/changelog work

+++ scripts/jira/getChangelog.js
@@ -1,3 +1,4 @@
+const fs = require('fs');
 const {executeJiraRequest, saveDataToFile, logger} = require("./utils");
 
 const CHANGELOG_URL = "https://ag-grid.atlassian.net/rest/api/2/search?jql=filter=11743+order+by+fixversion+desc";
@@ -9,6 +10,8 @@ const CHANGELOG_BAK_FILE = "/home/ceolter/jira_reports/changelog.bak.json";
 try {
     logger("Changelog Data Download Started", CHANGELOG_LOG);
     const changelogData = executeJiraRequest(CHANGELOG_URL);
+    logger("Changelog Data Downloaded", CHANGELOG_LOG);
+
     saveDataToFile(changelogData, CHANGELOG_TMP_FILE);
     logger("Changelog Data Downloaded", CHANGELOG_LOG);
 
@@ -16,7 +19,9 @@ try {
     if(changelogData.length > 1) {
         logger("Updating Existing Changelog Data", CHANGELOG_LOG);
         // if we've data switch it over to the "live" set, backing up the previous data set - just in case
-        fs.renameSync(CHANGELOG_FILE, CHANGELOG_BAK_FILE);
+        if(fs.existsSync(CHANGELOG_FILE)) {
+            fs.renameSync(CHANGELOG_FILE, CHANGELOG_BAK_FILE);
+        }
         fs.renameSync(CHANGELOG_TMP_FILE, CHANGELOG_FILE);
     }
